Add explicit casting

+++ datadriven/src/sgpp/datadriven/algorithm/DBMatDatabase.cpp
@@ -142,7 +142,7 @@ void DBMatDatabase::putDataMatrix(
       json::ListNode& level =
           dynamic_cast<json::ListNode&>(gridConfigEntry.addListAttr(keyGridLevel));
       for (uint64_t i = 0; i < gridConfig.levelVector_.size(); i++) {
-        level.addIdValue(gridConfig.levelVector_[i]);
+        level.addIdValue(static_cast<uint64_t>(gridConfig.levelVector_[i]));
       }
     } else {
       gridConfigEntry.addIDAttr(keyGridLevel, static_cast<int64_t>(gridConfig.level_));
